requirements.txt
absl-py==0.11.0
amqp==5.1.1
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
asgiref==3.6.0
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==20.3.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
<<<<<<< HEAD
billiard==3.6.4.0
bleach==3.3.0
blinker==1.4
braintree==4.16.0
Brlapi==0.7.0
Brotli==1.0.9
CacheControl==0.12.6
cachetools==4.1.1
cachy==0.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.14.5
chardet==3.0.4
chrome-gnome-shell==0.0.0
cleo==0.8.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clikit==0.6.2
colorama==0.4.3
command-not-found==0.3
configparser==5.0.1
crashtest==0.3.1
cryptography==2.8
cssselect2==0.6.0
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.2.18
django-bootstrap4==22.2
django-bootstrap5==23.1
django-embed-video==1.4.8
django-extensions==2.2.5
django-filter==21.1
django-heroku==0.3.1
django-localflavor==3.1
django-on-heroku==1.1.2
django-phone-field==1.8.1
django-phonenumber-field==7.0.0
django-rosetta==0.9.8
django-taggit==3.1.0
django-werkzeug-debugger-runserver==0.3.1
djangorestframework==3.13.1
docker-pycreds==0.4.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
flower==1.1.0
fonttools==4.34.4
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.11
google-api-python-client==1.7.11
google-auth==1.23.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
gunicorn==20.1.0
h5py==2.10.0
html5lib==1.1
httplib2==0.14.0
humanize==4.2.3
idna==2.8
image==1.5.33
imageio==2.9.0
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.7.0
jedi==0.18.0
jeepney==0.6.0
Jinja2==2.11.3
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kazam==1.4.5
Keras==2.4.3
Keras-Preprocessing==1.1.2
keyring==21.8.0
kiwisolver==1.3.1
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.3
MarkupSafe==1.1.0
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
mpmath==1.2.1
msgpack==1.0.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
netifaces==0.10.4
networkx==2.5
nose==1.3.7
notebook==6.2.0
notify2==0.3.1
numpy==1.18.5
oauth2client==4.1.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.4
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.8.1
pastel==0.2.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkginfo==1.7.0
poetry==1.1.5
poetry-core==1.0.2
polib==1.1.1
powerline-gitstatus==1.3.1
powerline-status==2.7
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.16
protobuf==3.14.0
psutil==5.7.3
psycopg==3.1.3
psycopg2==2.9.1
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
PyDrive==1.3.1
pydyf==0.2.0
pyenchant==2.0.0
pygame==2.0.1
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==2.6.0
pylev==1.3.0
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
PyPDF2==2.12.1
pyphen==0.12.0
pyRFC3339==1.1
pyrsistent==0.17.3
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-stdnum==1.17
python3-openid==3.2.0
pyttsx3==2.90
pytz==2022.1
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.0.3
qrcode==7.3
qtconsole==5.0.2
QtPy==1.9.0
redis==4.3.4
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
SecretStorage==3.3.1
Send2Trash==1.5.0
sentry-sdk==0.19.4
shellingham==1.4.0
shortuuid==1.0.1
simplejson==3.16.0
simplenote==2.1.4
six==1.14.0
smmap==3.0.4
social-auth-app-django==5.2.0
social-auth-core==4.4.1
=======
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-bootstrap==0.2.4
django-bootstrap4==22.1
django-embed-video==1.4.4
idna==3.3
Pillow==9.1.0
requests==2.27.1
>>>>>>> d32a961ec6d747a0d677f67036b53b44970dcf36
soupsieve==2.3.2.post1
SQLAlchemy==1.3.12
sqlparse==0.4.3
subprocess32==3.5.4
sympy==1.7.1
systemd-python==234
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.11.26
tinycss2==1.1.1
toml==0.10.2
tomlkit==0.7.0
tornado==6.1
traitlets==5.0.5
typing_extensions==4.4.0
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==0.6
urllib3==1.25.8
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.2.1
wadllib==1.3.3
wandb==0.10.11
watchdog==0.10.4
wcwidth==0.2.5
weasyprint==56.1
webencodings==0.5.1
Werkzeug==0.16.0
whitenoise==5.2.0
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
zopfli==0.2.1
